backend/requirements.txt
    Flask
    Flask-CORS
    stripe
    python-dotenv
    waitress
    gunicorn